split adaptive
