+++ src/adaptive/mod.rs
@@ -1,0 +1,18 @@
+//! This module contains the implementation of adaptive quadrature rules for both 1D and 2D.
+//!
+//! The entry point for the 1D adaptive quadrature algorithm is the `adaptive_quad` function, which
+//! resides in the `d1` module along with its options and results structs.
+//!
+//! The entry point
+//--------------------------------------------------------------------------------------------------
+
+//{{{ crate imports
+//}}}
+//{{{ std imports
+//}}}
+//{{{ dep imports
+//}}}
+//--------------------------------------------------------------------------------------------------
+
+pub mod d1;
+pub mod d2;
